backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.7.0
pydantic-settings==2.3.0
httpx==0.27.0
redis[hiredis]==5.0.0
python-jose[cryptography]==3.3.0
bcrypt==4.3.0
python-multipart==0.0.9
pillow==10.3.0
boto3==1.34.0
python-dotenv==1.0.1
